Project for GoldLibrary tests
just few tests added for now..
Main unit is .go, that calls to RunTests unit and that one does the DiscoverTests and so on.
All code to test is on .go, and asserts are in .pas units, this way we can test 'go mode'. If want to test 'oxygene mode' we can place the code in .pas units.